% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_description.R
\docType{data}
\name{amcel}
\alias{amcel}
\title{AMCEL plantation}
\format{
A `gpkg` file with 1 layer.
\describe{
   \item{amcel}{Polygon of the plantation.}
}
}
\description{
Polygon of the AMCEL - Amapá Florestal e Celulose S.A. plantation forests Amapá, Brazil.
}
\examples{
\donttest{
# check layers
famcel <- system.file("vector/amcel.gpkg", package="eprdados")
st_layers(famcel)
# load polygon
amcel <- st_read(famcel, layer ="amcel")
# take a look
library(mapview)
mapview(amcel)
}
}
