% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{amostra_mapbiomas_2020}
\alias{amostra_mapbiomas_2020}
\title{A sample of MapBiomas Collection 6 data}
\format{
`amostra_mapbiomas_2020`
is a raster with 8 rows and 9 columns:
\describe{
  \item{location}{Amapá, Brazil}
  \item{coord. ref}{SIRGAS 2000 / UTM zone 22N (EPSG:31976)}
  \item{year}{2020}
}
}
\source{
<https://brasil.mapbiomas.org/en/>
}
\usage{
amostra_mapbiomas_2020
}
\description{
A subset of data from MapBiomas Collection 6 classified land cover.
}
\examples{
# plot to see the raster
plot(rast(amostra_mapbiomas_2020))
}
\keyword{datasets}
