% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_description.R
\docType{data}
\name{class_5}
\alias{class_5}
\title{Marco Zero Campus in 5 classes.}
\format{
`class_5`
Is a classification of landcover.
The raster has 148 rows and 268 columns:
\describe{
  \item{location}{Amapá, Brazil}
  \item{coord. ref}{SIRGAS 2000 / UTM zone 22S (EPSG:31982)}
  \item{year}{2025}
}
}
\usage{
class_5
}
\description{
This is to do.
This is to do.
This is to do.
This is to do.
}
\details{
A classification of the Marco Zero Campus.
}
\examples{
# plot to see the raster cover classes
# internal raster data needs to be read
plot(terra::rast(class_5))
}
\keyword{datasets}
