% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{forest_cover_interviews}
\alias{forest_cover_interviews}
\title{Forest cover landscape metrics.}
\format{
`forest_cover_interviews`
is a dataframe, with:
\describe{
  \item{aid}{Property identifier.}
  \item{long}{Longitude. Decimal degrees (EPSG:4326).}
  \item{lat}{Latitude. Decimal degrees (EPSG:4326).}
  \item{lat}{Habitat_WWF. Terrestrial ecoregion. https://doi.org/10.1641/0006-3568(2001)051[0933:TEOTWA]2.0.CO;2}
  \item{metric}{Metric identifier from landscapemetrics.}
  \item{buff_dist_km}{Buffer distance. Radius in km.}
  \item{value_median}{Median value of landscape metric.}
}
}
\source{
<https://doi.org/10.1177/1940082920971747>
}
\usage{
forest_cover_interviews
}
\description{
Forest cover around rural property locations in Amapá, Brazil.
Includes landscape metrics for MapBiomas forest cover class.

Values for one metric \code{cpland} calculated using package  \href{https://r-spatialecology.github.io/landscapemetrics/}{landscapemetrics}.
\code{cpland} is the percentage of core area of class in relation to the total landscape area.
Landscape metric from \href{https://brasil.mapbiomas.org/}{MapBiomas} forest cover class.
Metrics are calculated around each property in buffers with radius of:
 0.125, 0.25, 0.5, 1, 2, 4, 8 and 16 km. To reduce influence of satellite
 image quality, the metric values are median values from three years:
 study year, year before and year after.
}
\details{
A subset of forest cover landscape metrics around properties.
}
\examples{
# check data. 106 properties and 8 buffer distances.
table(forest_cover_interviews$buff_dist_km)
}
\references{
Michalski, F., Boulhosa, R. L. P., Nascimento, Y. N. D., & Norris, D. (2020).
Rural wage-earners’ attitudes towards diverse wildlife groups differ between tropical ecoregions:
implications for forest and savanna conservation in the Brazilian Amazon.
Tropical Conservation Science, 13, https://doi.org/10.1177/1940082920971747.
}
\keyword{datasets}
