% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_description.R
\docType{data}
\name{marco_zero_2026_31982}
\alias{marco_zero_2026_31982}
\title{Marco Zero}
\format{
A `gpkg` file with 16 layers.
\describe{
   \item{marco_zero_2026_31982}{Polygons and points.}
}
}
\description{
Multiple layers of the Marco Zero Campus, Amapá, Brazil.
}
\examples{
\donttest{
# check layers
fmz <- system.file("vector/marco_zero_2026_31982.gpkg", package="eprdados")
st_layers(fmz)
# load polygon
campus <- st_read(fmz, layer ="campus")
# take a look
library(mapview)
mapview(campus)
}
}
