% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_description.R
\docType{data}
\name{rio_pontos}
\alias{rio_pontos}
\title{Sample points along rivers.}
\format{
`rio_pontos`
is sf POINT geometry with 52 points and 8 fields, including:
\describe{
  \item{subz_id}{Unique identifier.}
  \item{long}{Longitude. Decimal degrees.}
  \item{lat}{Latitude. Decimal degrees.}
  \item{impact}{Dam impact. 1 = yes, 0 = no.}
}
}
\source{
<https://fmichalski.wixsite.com/lecov>
}
\usage{
rio_pontos
}
\description{
The sample points are along centerline of rivers upstream of the
Cachoeira Caldeirão Dam.
A total of 52 points are regularly spaced at approximately 5 km intervals.
Used to develop before-after control-impact analysis of dam impacts on species including
Giant Otters (https://doi.org/10.7717/peerj.14287) and
Yellow-spotted River Turtles ( https://doi.org/10.1002/aqc.3775).
\describe{
  \item{location}{Amapá, Brazil.}
  \item{coord. ref}{WGS 84 (EPSG:4326).}
}
}
\details{
Sample points along Araguari and Falsino rivers.
}
\examples{
# plot to see the points
plot(rio_pontos["rio"])
}
\keyword{datasets}
