% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_description.R
\docType{data}
\name{rio_pontos_buff}
\alias{rio_pontos_buff}
\title{Buffers around sample point along rivers.}
\format{
`rio_pontos_buff`
is sf POLYGON geometry with 7 polygons and 9 fields, including:
\describe{
  \item{subz_id}{Unique identifier.}
  \item{long}{Longitude. Decimal degrees.}
  \item{lat}{Latitude. Decimal degrees.}
  \item{impact}{Dam impact. 1 = yes, 0 = no.}
  \item{raio}{Radius around point in meters.}
}
}
\source{
<https://fmichalski.wixsite.com/lecov>
}
\usage{
rio_pontos_buff
}
\description{
Circular buffers around a sample points along centerline of rivers upstream of the
Cachoeira Caldeirão Dam.
A total of 7 buffers with different radii.
\describe{
  \item{location}{Amapá, Brazil.}
  \item{coord. ref}{WGS 84 (EPSG:4326).}
}
}
\details{
Buffers around a sample point along the Araguari river.
}
\examples{
# plot to see the points
plot(rio_pontos_buff["raio"])
}
\keyword{datasets}
