% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{utm_cover_AP_rio_2020}
\alias{utm_cover_AP_rio_2020}
\title{A sample of MapBiomas Collection 6 data}
\format{
`utm_cover_AP_rio_2020`
is a raster with 5727 rows and 6059 columns:
\describe{
  \item{location}{Amapá, Brazil}
  \item{coord. ref}{SIRGAS 2000 / UTM zone 22N (EPSG:31976)}
  \item{year}{2020}
}
}
\source{
<https://brasil.mapbiomas.org/en/>
}
\usage{
utm_cover_AP_rio_2020
}
\description{
The sample covers an area upstream of the Cachoeira Caldeirão Dam.
This area has been studied to understand dam impacts on species including
Giant Otters (https://doi.org/10.7717/peerj.14287) and
Yellow-spotted River Turtles ( https://doi.org/10.1002/aqc.3775).
}
\details{
A subset of data from MapBiomas Collection 6 classified land cover.
}
\examples{
# plot to see the raster
plot(rast(utm_cover_AP_rio_2020))
}
\keyword{datasets}
